requirements.txt
beautifulsoup4
requests
openai
pinecone-client 
datasets
tqdm
pandas
mongod